--- src/main/resources/jdbs_task10/schema.sql
@@ -1,19 +1,0 @@
-CREATE
-DATABASE product;
-DROP
-DATABASE product;
-create table product
-(
-    id          BIGSERIAL PRIMARY KEY,
-    name        varchar(25),
-    cost        double precision,
-    promotional boolean
-);
-
-create table discountCard
-(
-    id   BIGSERIAL PRIMARY KEY,
-    sale int
-);
-
-
