--- src/main/resources/sql_task9/data.sql
@@ -1,89 +1,0 @@
-INSERT INTO firm    (pump_brand,    country)
-VALUES              ('DAB',         'Germany'),
-                    ('Caprari',     'Italian'),
-                    ('Xiro',        'Russian');
-
-INSERT INTO product (name,                  price,  engine_power,   amperage,   efficiency_pump,    producer_id)
-VALUES              ('SS6E8А',	            100,	5,	            10,         60,                 1),
-                    ('SS6E9А',	            200,	10,	            20,         70,	                1),
-                    ('SS6E10А',	            300,	15,	            30,	        80,	                1),
-                    ('S6E11А',	            400,	20,	            40,         90,	                1),
-                    ('SS6E12А',	            500,	25,	            50,	        95,                 1),
-                    ('SS6E13А',	            600,	30,	            60,	        99,                 1),
-                    ('SS6E14А',	            110,	16,	            24,	        55,	                1),
-                    ('SS6E15А',	            300,	15,	            30,	        80,	                1),
-                    ('SS6E16А',	            400,	20,	            40,         90,	                1),
-                    ('SS6E17А',	            500,	25,	            50,	        95,                 1),
-                    ('SS6E18А',	            600,	30,	            60,	        99,                 1),
-                    ('SS6E19А',	            110,	16,	            24,	        55,	                1),
-                    ('E8S64N-6/4K-V',	    115,	18,	            28,	        56,	                2),
-                    ('E8S65N-6/4K-V',	    120,	20,	            32,	        57,	                2),
-                    ('E8S66N-6/4K-V',	    125,	22,	            36,	        58,	                2),
-                    ('E8S67N-6/4K-V',	    130,	24,	            40,	        59,	                2),
-                    ('E8S68N-6/4K-V',	    135,	26,	            44,	        60,	                2),
-                    ('E8S69N-6/4K-V',	    140,	28,	            48,         61,	                2),
-                    ('E8S70N-6/4K-V',	    145,	30,	            52,         62,	                2),
-                    ('SPI 07.95-5-A1',	    50,	    16,	            48,	        35,	                3),
-                    ('SPI 07.95-5-A2',	    59,	    36,	            64,         37,	                3),
-                    ('SPI 07.95-5-A3',	    68,	    56,	            80,	        39,	                3),
-                    ('SPI 07.95-5-A4',	    77,	    76,	            96,	        41,	                3),
-                    ('SPI 07.95-5-A5',	    86,	    96,	            112,        43,	                3),
-                    ('SPI 07.95-5-A6', 	    95,	    116,	        128,        45,	                3),
-                    ('SPI 07.95-5-A7',	    104,	136,	        144,        47,	                3),
-                    ('SPI 07.95-5-A8',	    113,	156,	        160,        49,	                3);
-
-INSERT INTO company ( name,              city,       adress,             tel_code,      tel_number    )
-VALUES              ('Gomelvodokanal',   'Gomel',    'Lenina 35',        375232,        445588      ),
-                    ('Brestvodokanl',    'Brest',    'Sovetskay 76',     375232,        663352      ),
-                    ('Vitebskvodokanal', 'Vitebsk',  'International 78', 375232,        225365      ),
-                    ('Vitebskvodokanal', 'Vitebsk',  'International 78', 375232,        225365      );
-
-INSERT INTO purchases   (company_id,    product_id,     count)
-VALUES                  (1,             1,              5),
-                        (1,             2,              26),
-                        (2,             3,              5),
-                        (2,             4,              2),
-                        (3,             5,              55),
-                        (3,             6,              5),
-                        (4,             7,              4),
-                        (4,             8,              7),
-                        (1,             9,              33),
-                        (1,             10,             22),
-                        (2,             11,             44),
-                        (2,             12,             12),
-                        (3,             13,             13),
-                        (3,             14,             10),
-                        (4,             15,             9),
-                        (4,             16,             8),
-                        (1,             17,             7),
-                        (1,             18,             6),
-                        (2,             19,             2),
-                        (2,             20,             52),
-                        (3,             21,             132),
-                        (3,             22,             5),
-                        (4,             23,             8),
-                        (4,             24,             8),
-                        (1,             25,             8),
-                        (2,             26,             4),
-                        (1,             9,              33),
-                        (1,             1,             22),
-                        (2,             1,             44),
-                        (2,             3,             12),
-                        (3,             3,             13),
-                        (3,             1,             10),
-                        (4,             1,             9),
-                        (4,             10,             8),
-                        (1,             10,             7),
-                        (1,             11,             6),
-                        (2,             11,             2),
-                        (2,             10,             52),
-                        (3,             2,             132),
-                        (3,             1,             5),
-                        (4,             2,             8),
-                        (4,             1,             8),
-                        (1,             25,             8),
-                        (2,             26,             4);
-drop table  company;
-drop table  firm;
-drop table  product;
-drop table  purchases;
